+++ src/components/Responsive/index.tsx
@@ -1,13 +1,24 @@
 import * as React from 'react';
-import { getMediaQueries } from '../../utils/displayTo';
-import { PreciseTheme, ScreenSize, ScreenSizeList } from '../../common';
+import {
+  getWidthBreakpointsQuery,
+  getScreenSizeBreakpoints,
+  WidthBreakpoints,
+  getMediaQueries,
+} from '../../utils/displayTo';
+import { PreciseTheme, ScreenSize, Breakpoints } from '../../common';
 import { light, breakpoints } from '../../themes';
+
+export type ScreenSizeCondition = 'equal' | 'from' | 'upTo';
 
 export interface ResponsiveProps {
   /**
    * Sets the screen size for which to show the children.
    */
   screenSize?: ScreenSize;
+  /*
+   * Sets the screen size check condition.
+   */
+  screenSizeCondition?: ScreenSizeCondition;
   /**
    * Sets the render callback. When screen size changes it's called with a new one.
    */
@@ -28,26 +39,52 @@ export interface ResponsiveState {
 
 export class Responsive extends React.Component<ResponsiveProps, ResponsiveState> {
   private mediaQueriesToListen: { [size: string]: MediaQueryList } = {};
-  private allMediaQueries: { [size: string]: string } = {};
 
   constructor(props: ResponsiveProps) {
     super(props);
 
-    const { theme = light, screenSize, render } = this.props;
-    this.allMediaQueries = getMediaQueries({ ...breakpoints, ...theme.breakpoints });
-    if (typeof render !== 'function' && (!screenSize || Object.keys(this.allMediaQueries).indexOf(screenSize) === -1)) {
-      console.error('`render` callback method or valid `screenSize` should be defined');
-    } else {
-      const screenSizesListenTo = screenSize ? [screenSize] : [...ScreenSizeList];
+    const { theme = light, screenSize, screenSizeCondition = 'equal', render } = this.props;
+    const themedBreakpoints = { ...breakpoints, ...theme.breakpoints };
 
-      for (const screenSizeListenTo of screenSizesListenTo) {
-        this.mediaQueriesToListen[screenSizeListenTo] = window.matchMedia(this.allMediaQueries[screenSizeListenTo]);
+    if (screenSize) {
+      const screenSizeBreakpoints = this.getBreakpoints(themedBreakpoints, screenSize, screenSizeCondition);
+      if (screenSizeBreakpoints) {
+        this.mediaQueriesToListen[screenSize] = window.matchMedia(getWidthBreakpointsQuery(screenSizeBreakpoints));
       }
+    } else {
+      const allMediaQueries = getMediaQueries(themedBreakpoints);
+      for (const screenSize in allMediaQueries) {
+        this.mediaQueriesToListen[screenSize] = window.matchMedia(allMediaQueries[screenSize]);
+      }
+    }
+
+    if (typeof render !== 'function' && !Object.keys(this.mediaQueriesToListen).length) {
+      console.error('`render` callback method or valid `screenSize` should be defined');
     }
 
     this.state = {
       matchesTo: undefined,
     };
+  }
+
+  private getBreakpoints(breakpoints: Breakpoints, screenSize: ScreenSize, screenSizeCondition: ScreenSizeCondition) {
+    const screenSizeBreakpoints = getScreenSizeBreakpoints(screenSize, breakpoints);
+    if (!screenSizeBreakpoints) {
+      console.error('Invalid screen size');
+      return;
+    }
+
+    switch (screenSizeCondition) {
+      case 'equal':
+        return { ...screenSizeBreakpoints };
+      case 'from':
+        return { ...screenSizeBreakpoints, max: undefined };
+      case 'upTo':
+        return { ...screenSizeBreakpoints, min: undefined };
+      default:
+        console.error('Invalid compare condition');
+        return;
+    }
   }
 
   private currentlyMatches() {
@@ -69,11 +106,7 @@ export class Responsive extends React.Component<ResponsiveProps, ResponsiveState
   };
 
   componentDidMount() {
-    const { screenSize } = this.props;
-    const screenSizesListenTo = screenSize ? [screenSize] : [...ScreenSizeList];
-
-    for (const screenSizeListenTo of screenSizesListenTo) {
-      this.mediaQueriesToListen[screenSizeListenTo] = window.matchMedia(this.allMediaQueries[screenSizeListenTo]);
+    for (const screenSizeListenTo in this.mediaQueriesToListen) {
       this.mediaQueriesToListen[screenSizeListenTo].addListener(this.updateMatches);
     }